# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "correlation_id",
          "details"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          },
          "schema_version": {
            "const": "v1",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "IdentityResponse": {
        "additionalProperties": false,
        "properties": {
          "principal": {
            "$ref": "#/components/schemas/PrincipalIdentityResponse"
          },
          "schema_version": {
            "const": "v1",
            "title": "Schema Version",
            "type": "string"
          },
          "tenant": {
            "$ref": "#/components/schemas/TenantIdentityResponse"
          }
        },
        "required": [
          "schema_version",
          "tenant",
          "principal"
        ],
        "title": "IdentityResponse",
        "type": "object"
      },
      "MissionCreate": {
        "additionalProperties": false,
        "properties": {
          "audience": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Audience",
            "type": "string"
          },
          "budget_cents": {
            "default": 0,
            "maximum": 100000000.0,
            "minimum": 0.0,
            "title": "Budget Cents",
            "type": "integer"
          },
          "campaign_goal": {
            "default": "awareness",
            "maxLength": 120,
            "minLength": 1,
            "title": "Campaign Goal",
            "type": "string"
          },
          "objective": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Objective",
            "type": "string"
          },
          "platforms": {
            "items": {
              "$ref": "#/components/schemas/Platform"
            },
            "maxItems": 4,
            "minItems": 1,
            "title": "Platforms",
            "type": "array"
          },
          "schema_version": {
            "const": "v1",
            "title": "Schema Version",
            "type": "string"
          },
          "source_asset": {
            "default": "sandbox://brief/no-external-asset",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Source Asset",
            "type": "string"
          },
          "title": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "title",
          "objective",
          "audience",
          "platforms"
        ],
        "title": "MissionCreate",
        "type": "object"
      },
      "MissionResponse": {
        "additionalProperties": false,
        "properties": {
          "audience": {
            "title": "Audience",
            "type": "string"
          },
          "budget_cents": {
            "title": "Budget Cents",
            "type": "integer"
          },
          "campaign_goal": {
            "title": "Campaign Goal",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "mission_id": {
            "title": "Mission Id",
            "type": "string"
          },
          "objective": {
            "title": "Objective",
            "type": "string"
          },
          "platforms": {
            "items": {
              "$ref": "#/components/schemas/Platform"
            },
            "title": "Platforms",
            "type": "array"
          },
          "schema_version": {
            "const": "v1",
            "title": "Schema Version",
            "type": "string"
          },
          "source_asset": {
            "title": "Source Asset",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "schema_version",
          "mission_id",
          "tenant_id",
          "created_by",
          "title",
          "objective",
          "audience",
          "platforms",
          "budget_cents",
          "source_asset",
          "campaign_goal",
          "created_at",
          "version"
        ],
        "title": "MissionResponse",
        "type": "object"
      },
      "Platform": {
        "enum": [
          "x",
          "facebook",
          "tiktok",
          "instagram"
        ],
        "title": "Platform",
        "type": "string"
      },
      "PrincipalIdentityResponse": {
        "additionalProperties": false,
        "properties": {
          "auth_mode": {
            "title": "Auth Mode",
            "type": "string"
          },
          "principal_id": {
            "title": "Principal Id",
            "type": "string"
          },
          "schema_version": {
            "const": "v1",
            "title": "Schema Version",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "tenant_id",
          "principal_id",
          "auth_mode"
        ],
        "title": "PrincipalIdentityResponse",
        "type": "object"
      },
...
      "TenantIdentityResponse": {
        "additionalProperties": false,
        "properties": {
          "schema_version": {
            "const": "v1",
            "title": "Schema Version",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "tenant_id"
        ],
        "title": "TenantIdentityResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Versioned, tenant-scoped control plane. External providers remain sandbox-only and approvals never imply publication or ad spend.",
    "title": "AI-Native Content Agency Control Plane",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/identity": {
      "get": {
        "operationId": "get_identity_endpoint_api_v1_identity_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Tenant-ID",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$",
              "title": "X-Tenant-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Principal-ID",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$",
              "title": "X-Principal-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Validation Failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get Identity Endpoint",
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/missions": {
      "post": {
        "operationId": "create_mission_endpoint_api_v1_missions_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Tenant-ID",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$",
              "title": "X-Tenant-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Principal-ID",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$",
              "title": "X-Principal-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MissionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Body Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Validation Failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database or Dependency Unavailable"
          }
        },
        "summary": "Create Mission Endpoint",
        "tags": [
          "missions"
        ]
      }
    },
...
  }
}