# src/api/contracts.ts
/**
 * GENERATED from backend/openapi.json by scripts/generate_ts_contracts.py.
 * Do not edit this file directly; update the Pydantic contract and regenerate it.
 */

export const SCHEMA_VERSION = "v1" as const;
export const GREENLIGHT_POLICY_VERSION = "greenlight.v1" as const;

export type SchemaVersion = typeof SCHEMA_VERSION;
export type GreenlightPolicyVersion = typeof GREENLIGHT_POLICY_VERSION;

export type AgentRole = "ceo" | "research" | "strategist" | "growth" | "writer" | "media" | "risk" | "publisher";

export type AgentStatus = "standby" | "processing" | "ready" | "waiting_greenlight" | "blocked" | "attention";

export interface ApprovalCreate {
  artifact_manifest_hash: string;
  decision: ApprovalDecision;
  note?: string;
  policy_version: GreenlightPolicyVersion;
  reviewer: string;
  schema_version: SchemaVersion;
}

export type ApprovalDecision = "approved" | "rejected";

export interface ApprovalResponse {
  approval_id: string;
  artifact_manifest_hash: string;
  decided_at: string;
  decision: ApprovalDecision;
  idempotency_key: string;
  note: string;
  policy_version: GreenlightPolicyVersion;
  principal_id: string;
  reviewer: string;
  schema_version: SchemaVersion;
}

export interface ArtifactResponse {
  artifact_id: string;
  created_at: string;
  created_by: AgentRole;
  evidence_ids: string[];
  kind: string;
  ordinal: number;
  payload: Record<string, unknown>;
  schema_version: SchemaVersion;
  title: string;
}

export interface AuditEventResponse {
  action: string;
  audit_id: string;
  correlation_id: string;
  occurred_at: string;
  payload: Record<string, unknown>;
  principal_id: string;
  schema_version: SchemaVersion;
}

export interface ErrorBody {
  code: string;
  correlation_id: string;
  details: Record<string, unknown>;
  message: string;
}

export interface ErrorResponse {
  error: ErrorBody;
  schema_version: SchemaVersion;
}

export interface HealthResponse {
  schema_version: SchemaVersion;
  status: "ok" | "ready";
}

export interface IdentityResponse {
  principal: PrincipalIdentityResponse;
  schema_version: SchemaVersion;
  tenant: TenantIdentityResponse;
}

export interface MissionCreate {
  audience: string;
  budget_cents?: number;
  campaign_goal?: string;
  objective: string;
  platforms: Platform[];
  schema_version: SchemaVersion;
  source_asset?: string;
  title: string;
}

export interface MissionResponse {
  audience: string;
  budget_cents: number;
  campaign_goal: string;
  created_at: string;
  created_by: string;
  mission_id: string;
  objective: string;
  platforms: Platform[];
  schema_version: SchemaVersion;
  source_asset: string;
  tenant_id: string;
  title: string;
  version: number;
}

export type Platform = "x" | "facebook" | "tiktok" | "instagram";

export interface PrincipalIdentityResponse {
  auth_mode: string;
  principal_id: string;
  schema_version: SchemaVersion;
  tenant_id: string;
}

export interface RunEventResponse {
  action: string;
  artifact_ids: string[];
  detail: string;
  event_id: string;
  evidence_ids: string[];
  role: AgentRole;
  schema_version: SchemaVersion;
  sequence: number;
  status: AgentStatus;
  timestamp: string;
}

export interface RunResponse {
  approval: ApprovalResponse | null;
  artifact_manifest_hash: string;
  artifacts: ArtifactResponse[];
  audit_events: AuditEventResponse[];
  completed_at: string | null;
  events: RunEventResponse[];
  evidence: ToolEvidenceResponse[];
  external_side_effects: boolean;
  mission_id: string;
  policy_version: GreenlightPolicyVersion;
  run_id: string;
  schema_version: SchemaVersion;
  started_at: string;
  status: RunStatus;
  steps: RunStepResponse[];
  tenant_id: string;
  version: number;
}

export interface RunStart {
  schema_version: SchemaVersion;
}

export type RunStatus = "running" | "awaiting_greenlight" | "completed" | "rejected" | "failed";

export interface RunStepResponse {
  detail: string;
  progress: number;
  role: AgentRole;
  schema_version: SchemaVersion;
  sequence: number;
  status: AgentStatus;
  step_id: string;
  updated_at: string;
}

export interface TenantIdentityResponse {
  schema_version: SchemaVersion;
  tenant_id: string;
}

export interface ToolEvidenceResponse {
  created_at: string;
  evidence_id: string;
  operation: string;
  payload: Record<string, unknown>;
  references: string[];
  sandbox: boolean;
  schema_version: SchemaVersion;
  summary: string;
  tool: string;
}